refactor backup pickup

+++ service/src/main/kotlin/fi/espoo/evaka/backuppickup/BackupPickupController.kt
@@ -5,11 +5,12 @@
 package fi.espoo.evaka.backuppickup
 
 import fi.espoo.evaka.Audit
-import fi.espoo.evaka.shared.auth.AccessControlList
+import fi.espoo.evaka.shared.BackupPickupId
 import fi.espoo.evaka.shared.auth.AuthenticatedUser
-import fi.espoo.evaka.shared.auth.UserRole
 import fi.espoo.evaka.shared.db.Database
-import fi.espoo.evaka.shared.domain.NotFound
+import fi.espoo.evaka.shared.db.updateExactlyOne
+import fi.espoo.evaka.shared.security.AccessControl
+import fi.espoo.evaka.shared.security.Action
 import org.jdbi.v3.core.kotlin.mapTo
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.DeleteMapping
@@ -22,7 +23,23 @@ import org.springframework.web.bind.annotation.RestController
 import java.util.UUID
 
 @RestController
-class BackupPickupController(private val acl: AccessControlList) {
+class BackupPickupController(private val accessControl: AccessControl) {
+    @PostMapping("/children/{childId}/backup-pickups")
+    fun createForChild(
+        db: Database.Connection,
+        user: AuthenticatedUser,
+        @PathVariable("childId") childId: UUID,
+        @RequestBody body: ChildBackupPickupContent
+    ): ResponseEntity<ChildBackupPickupCreateResponse> {
+        Audit.ChildBackupPickupCreate.log(targetId = childId)
+        accessControl.requirePermissionFor(user, Action.Child.CREATE_BACKUP_PICKUP, childId)
+        return db.transaction { tx ->
+            tx.createBackupPickup(childId, body)
+        }.let {
+            ResponseEntity.ok(ChildBackupPickupCreateResponse(it))
+        }
+    }
+
     @GetMapping("/children/{childId}/backup-pickups")
     fun getForChild(
         db: Database.Connection,
@@ -30,58 +47,37 @@ class BackupPickupController(private val acl: AccessControlList) {
         @PathVariable("childId") childId: UUID
     ): ResponseEntity<List<ChildBackupPickup>> {
         Audit.ChildBackupPickupRead.log(targetId = childId)
-        acl.getRolesForChild(user, childId).requireOneOfRoles(UserRole.ADMIN, UserRole.UNIT_SUPERVISOR, UserRole.STAFF, UserRole.SPECIAL_EDUCATION_TEACHER)
+        accessControl.requirePermissionFor(user, Action.Child.READ_BACKUP_PICKUP, childId)
 
         return db.transaction { tx ->
             tx.getBackupPickupsForChild(childId)
         }.let { ResponseEntity.ok(it) }
     }
 
-    @PostMapping("/children/{childId}/backup-pickups")
-    fun createForChild(
+    @PutMapping("/backup-pickups/{id}")
+    fun update(
         db: Database.Connection,
         user: AuthenticatedUser,
-        @PathVariable("childId") childId: UUID,
-        @RequestBody body: NewChildBackupPickup
-    ): ResponseEntity<ChildBackupPickupCreateResponse> {
-        Audit.ChildBackupPickupCreate.log(targetId = childId)
-        acl.getRolesForChild(user, childId).requireOneOfRoles(UserRole.ADMIN, UserRole.UNIT_SUPERVISOR, UserRole.STAFF)
-        return db.transaction { tx ->
-            tx.createBackupPickup(
-                body
-            )
-        }.let {
-            ResponseEntity.ok(ChildBackupPickupCreateResponse(it))
-        }
-    }
-
-    @PutMapping("/children/{childId}/backup-pickups")
-    fun updateForChild(
-        db: Database.Connection,
-        user: AuthenticatedUser,
-        @PathVariable("childId") childId: UUID,
-        @RequestBody body: ChildBackupPickup
+        @PathVariable("id") id: BackupPickupId,
+        @RequestBody body: ChildBackupPickupContent
     ): ResponseEntity<Unit> {
-        Audit.ChildBackupPickupUpdate.log(targetId = childId)
-        acl.getRolesForChild(user, childId).requireOneOfRoles(UserRole.ADMIN, UserRole.UNIT_SUPERVISOR, UserRole.STAFF)
+        Audit.ChildBackupPickupUpdate.log(targetId = id)
+        accessControl.requirePermissionFor(user, Action.BackupPickup.UPDATE, id)
         return db.transaction { tx ->
-            tx.updateBackupPickup(
-                body
-            )
+            tx.updateBackupPickup(id, body)
         }.let {
             ResponseEntity.noContent().build()
         }
     }
 
-    @DeleteMapping("/children/{childId}/backup-pickups/{id}")
+    @DeleteMapping("/backup-pickups/{id}")
     fun delete(
         db: Database.Connection,
         user: AuthenticatedUser,
-        @PathVariable("id") id: UUID,
-        @PathVariable("childId") childId: UUID
+        @PathVariable("id") id: BackupPickupId
     ): ResponseEntity<Unit> {
-        Audit.ChildBackupPickupDelete.log(targetId = childId)
-        acl.getRolesForChild(user, childId).requireOneOfRoles(UserRole.ADMIN, UserRole.UNIT_SUPERVISOR, UserRole.STAFF)
+        Audit.ChildBackupPickupDelete.log(targetId = id)
+        accessControl.requirePermissionFor(user, Action.BackupPickup.DELETE, id)
         db.transaction { tx ->
             tx.deleteBackupPickup(id)
         }
@@ -89,69 +85,68 @@ class BackupPickupController(private val acl: AccessControlList) {
     }
 }
 
-fun Database.Transaction.deleteBackupPickup(
-    id: UUID
-) {
-    val deleted = this.createUpdate("DELETE FROM backup_pickup WHERE id = :id")
-        .bind("id", id)
-        .execute()
-
-    if (deleted == 0) throw NotFound("No backup pickup $id found")
-}
-
-fun Database.Read.getBackupPickupsForChild(childId: UUID): List<ChildBackupPickup> {
-    return createQuery("SELECT * FROM backup_pickup WHERE child_Id = :id")
-        .bind("id", childId)
-        .mapTo<ChildBackupPickup>()
-        .list()
-}
-
 fun Database.Transaction.createBackupPickup(
-    backupPickup: NewChildBackupPickup
-): UUID {
+    childId: UUID,
+    data: ChildBackupPickupContent
+): BackupPickupId {
     // language=sql
     val sql = """
         INSERT INTO backup_pickup (child_id, name, phone)
         VALUES (:childId, :name, :phone)
         RETURNING id
     """.trimIndent()
+
     return this.createQuery(sql)
-        .bind("childId", backupPickup.childId)
-        .bind("name", backupPickup.name)
-        .bind("phone", backupPickup.phone)
-        .mapTo<UUID>()
-        .first()
+        .bind("childId", childId)
+        .bind("name", data.name)
+        .bind("phone", data.phone)
+        .mapTo<BackupPickupId>()
+        .one()
+}
+
+fun Database.Read.getBackupPickupsForChild(childId: UUID): List<ChildBackupPickup> {
+    return createQuery("SELECT id, child_id, name, phone FROM backup_pickup WHERE child_Id = :id")
+        .bind("id", childId)
+        .mapTo<ChildBackupPickup>()
+        .list()
 }
 
 fun Database.Transaction.updateBackupPickup(
-    backupPickup: ChildBackupPickup
+    id: BackupPickupId,
+    data: ChildBackupPickupContent
 ) {
     // language=sql
     val sql = """
         UPDATE backup_pickup
-        SET
-            name = :name,
-            phone = :phone
+        SET name = :name, phone = :phone
         WHERE id  = :id
     """.trimIndent()
-    val updated = this.createUpdate(sql)
-        .bind("id", backupPickup.id)
-        .bind("name", backupPickup.name)
-        .bind("phone", backupPickup.phone)
-        .execute()
 
-    if (updated == 0) throw NotFound("No backup pickup ${backupPickup.id} found")
+    this.createUpdate(sql)
+        .bind("id", id)
+        .bind("name", data.name)
+        .bind("phone", data.phone)
+        .updateExactlyOne()
 }
 
-data class NewChildBackupPickup(
-    val childId: UUID,
+fun Database.Transaction.deleteBackupPickup(
+    id: BackupPickupId
+) {
+    this.createUpdate("DELETE FROM backup_pickup WHERE id = :id")
+        .bind("id", id)
+        .updateExactlyOne()
+}
+
+data class ChildBackupPickupContent(
     val name: String,
     val phone: String
 )
+
 data class ChildBackupPickup(
-    val id: UUID,
+    val id: BackupPickupId,
     val childId: UUID,
     val name: String,
     val phone: String
 )
-data class ChildBackupPickupCreateResponse(val id: UUID)
+
+data class ChildBackupPickupCreateResponse(val id: BackupPickupId)
